// src/app/shared/interfaces/film-model.ts
export interface Film {
  adult?: boolean;
  backdrop_path: string;
  genre_ids: Array<number>;
  id: number;
  original_language: string;
  original_title: string;
  overview: string;
  popularity: number;
  poster_path: string;
  release_date: string;
  title: string;
  video: boolean;
  vote_average: number;
  vote_count: number;
  isWatched?: boolean;
}

export interface FilmsApiResponse {
  page: number;
  results: Film[];
  total_pages: number;
  total_results: number;
}